b2b_backend/requirements.txt
Django==1.11
django-appconf==1.0.2
django-cors-headers==2.3.0
django-imagekit==4.0.2
django-libs==2.0.0
django-object-actions==1.0.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
pilkit==2.0
Pillow==5.2.0
psycopg2==2.7.4
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
